:recycle: refactor(tests): separate E2E conftest from root and split CI test steps
